Split conversion into internal and e2e tests
